+++ Ejercicio5.java
@@ -1,0 +1,11 @@
+import tads.Grafo;
+
+public class Ejercicio5 {
+    public static void main(String[] args) {
+        Grafo<Integer> g = Grafo.convertFromText();
+        ejercicio5(g, 1);
+    }
+
+    public static void ejercicio5(Grafo<Integer> g, Integer vertice) {
+    }
+}
